% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_targets_pipeline.R
\name{build_targets_pipeline}
\alias{build_targets_pipeline}
\title{Dynamically build the targets pipeline for the project.}
\usage{
build_targets_pipeline(
  targets_notebook_name = "_target_notebook.Rmd",
  targets_file_name = "_targets.R",
  pkg_name = "KiribatiBiomarkers",
  vignette_name = "pipeline-definition.Rmd",
  overwrite = FALSE
)
}
\arguments{
\item{targets_notebook_name}{Character. The name of the local targets notebook file to create. Default is \code{"_target_notebook.Rmd"}.}

\item{targets_file_name}{Character. The name of the local targets file to create. Default is \code{"_targets.R"}.}

\item{pkg_name}{Character. This package, \code{"NepalClimateHealth"}.}

\item{vignette_name}{Character. The name of the vignette in the package that defines the targets notebook. Default is \code{"pipeline-definition"}.}

\item{overwrite}{Logical. Whether to overwrite existing files. Default is \code{FALSE}.}
}
\value{
Invisible \code{NULL}. The function is called for its side effects of creating files.
}
\description{
Dynamically build the targets pipeline for the project.
}
\details{
This function sets up the targets pipeline by defining a local \verb{_targets.R} file
and using \code{tar_tangle()} to create the necessary targets from an Rmd file.
The Rmd file is, itself, the vignette \code{vignette("pipeline-definition", package = "NepalClimateHealth")}.
Note that the function checks for the presence of required packages and alerts the user if any are missing,
because they are necessary to build the pipeline from the vignette locally.
}
\examples{
#build_targets_pipeline()
}
