% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KiribatiBiomarkers-package.R
\docType{package}
\name{KiribatiBiomarkers-package}
\alias{KiribatiBiomarkers}
\alias{KiribatiBiomarkers-package}
\title{KiribatiBiomarkers: Kiribati Biomarkers Data Distribution Package}
\description{
This package provides access to biomarker data collected in Kiribati.
}
\author{
\strong{Maintainer}: Tinashe Tapera \email{ttapera@hsph.harvard.edu}

Other contributors:
\itemize{
  \item Jessica Mason \email{jzm@hsph.harvard.edu} [compiler]
  \item Christopher Golden \email{golden@hsph.harvard.edu} [copyright holder]
}

}
\keyword{internal}
